requirements.txt
py7zr==0.21.0
pandas==2.1.4
google-cloud-storage==2.14.0